requirements.txt
# 機械学習
xgboost==2.0.0
lightgbm==4.0.0
scikit-learn==1.3.0
prophet==1.1.4
statsmodels==0.14.0

# 深層学習（LSTM用、オプショナル - CPU環境では不要）
# tensorflow==2.13.0  # 必要に応じてコメントアウトを解除

# 解釈可能性
shap==0.42.1
lime==0.2.0.1

# データ処理
pandas==2.0.3
numpy==1.24.3

# 時系列処理
tsfresh==0.20.1

# 可視化
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# ダッシュボード
streamlit==1.25.0

# ハイパーパラメータチューニング
optuna==3.3.0

# データ取得
requests==2.31.0
beautifulsoup4==4.12.0

# その他
pyyaml==6.0.1
jupyter==1.0.0
tqdm==4.66.1
